[fix] 플레이어 공격 버그 수정
- 트리거 script를 hammer 쪽으로 변경
- 보스 트리거 오류 수정
