+++ Assets/Scripts/HammerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b54752f261e92424386bd06264fda0e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
